+++ rsass/tests/spec/non_conformant/basic/t14_imports.rs
@@ -10,7 +10,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // wrong result
 fn test() {
     assert_eq!(
         runner().ok("@import \"a.scss\";\n\